\name{invDiag}
\alias{invDiag}
\title{
Matrix Diagonals with a useless(?) twist
}
\description{
Construct a diagonal matrix, with a pi/2 rotation.
}
\usage{
invDiag(x)
}
\arguments{
  \item{x}{
scalar integer. Returns a square matrix of size given by the scalar.
}
}
\details{
Probably useless.
}
\value{
A square matrix.
}
\author{
Antoine Filipovic Pierucci
}
\seealso{
\code{\link{diag}}.
}
\examples{
invDiag(5)
}
\keyword{ misc }
