\name{timeVar}
\alias{timeVar}
\title{
Generate Time-Dependant Variables
}
\description{
Generate time-dependant variables for survival models.
}
\usage{
timeVar(x)
}
\arguments{
\item{x}{rien}
}
\details{
Not implemented yet.
}
\value{
Nothing yet.
}
\author{
Antoine Filipovic Pierucci
}
\seealso{
\code{\link{coxph}} (package survival).
}
\examples{
timeVar()

}
\keyword{ misc }
